LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_eval-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
Scripts injected via `eval` are not allowed with `strict-dynamic` without `unsafe-eval`.


Harness Error (TIMEOUT), message = null

TIMEOUT Script injected via `eval` is not allowed with `strict-dynamic` without `unsafe-eval`. Test timed out
